+++ app/action/action-category.ts
@@ -1,15 +1,14 @@
 "use server";
 
+import prisma from "../components/lib/prisma";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
-import prisma from "../components/lib/prisma";
-import { supabase } from "../components/util/supabase";
 import { z } from "zod";
-import { FileSaveUtils } from "../components/lib/FileSaveUtils";
+import { supabase } from "../components/util/supabase";
+import { fileSaveUtils } from "../lib/fileSaveUtils";
 import { validateFile } from "../components/lib/ValidateFile";
 import { revalidatePostsAndCategories } from "../components/lib/revalidatePostsAndCategories";
 import { getCategory } from "../components/lib/BlogServiceUnique";
-import { getCurrentUserRole } from "../components/lib/getCurrentUser";
 import { checkUserRole } from "../components/lib/checkUserRole";
 
 type FormState = {
@@ -108,7 +107,7 @@ export const addCategory = async (state: FormState, data: FormData) => {
         return errors;
       }
 
-      const { fileUrl, fileName } = await FileSaveUtils(image);
+      const { fileUrl, fileName } = await fileSaveUtils(image);
       const createdImage = await prisma.postImage.create({
         data: {
           name: fileName,
@@ -280,7 +279,7 @@ export const updateCategory = async (
         return errors;
       }
 
-      const { fileUrl, fileName } = await FileSaveUtils(image);
+      const { fileUrl, fileName } = await fileSaveUtils(image);
 
       const createdImage = await prisma.postImage.create({
         data: {
