+++ app/action/action-post.ts
@@ -5,7 +5,7 @@ import { redirect } from "next/navigation";
 import prisma from "../components/lib/prisma";
 import { supabase } from "../components/util/supabase";
 import { z } from "zod";
-import { FileSaveUtils } from "../components/lib/FileSaveUtils";
+import { fileSaveUtils } from "../lib/fileSaveUtils";
 import { validateFile } from "../components/lib/ValidateFile";
 import { revalidatePostsAndCategories } from "../components/lib/revalidatePostsAndCategories";
 import { getPost } from "../components/lib/BlogServiceUnique";
@@ -112,7 +112,7 @@ export const addPost = async (state: FormState, data: FormData) => {
         return errors;
       }
 
-      const { fileUrl, fileName } = await FileSaveUtils(image);
+      const { fileUrl, fileName } = await fileSaveUtils(image);
       const createdImage = await prisma.postImage.create({
         data: {
           name: fileName,
@@ -267,7 +267,7 @@ export const updatePost = async (
         return errors;
       }
 
-      const { fileUrl, fileName } = await FileSaveUtils(image);
+      const { fileUrl, fileName } = await fileSaveUtils(image);
       const createdImage = await prisma.postImage.create({
         data: {
           name: fileName,
