+++ app/action/action-postImage.ts
@@ -6,7 +6,7 @@ import prisma from "../components/lib/prisma";
 import { supabase } from "../components/util/supabase";
 import { z } from "zod";
 import { promises as fsPromises } from "fs";
-import { FileSaveUtils } from "../components/lib/FileSaveUtils";
+import { fileSaveUtils } from "../lib/fileSaveUtils";
 import { validateFile } from "../components/lib/ValidateFile";
 import { getPostImage } from "../components/lib/BlogServiceUnique";
 import { checkUserRole } from "../components/lib/checkUserRole";
@@ -70,7 +70,7 @@ export const addPostImage = async (state: FormState, data: FormData) => {
       return errors;
     }
 
-    const { fileUrl, fileName } = await FileSaveUtils(image);
+    const { fileUrl, fileName } = await fileSaveUtils(image);
 
     await prisma.postImage.create({
       data: {
@@ -191,7 +191,7 @@ export const updatePostImage = async (
     }
 
     try {
-      const { fileUrl, fileName } = await FileSaveUtils(image);
+      const { fileUrl, fileName } = await fileSaveUtils(image);
 
       await prisma.postImage.update({
         where: {
